+++ bundles/remote_services/discovery_common/src/discovery_activator.c
@@ -96,7 +96,7 @@ celix_status_t bundleActivator_start(void * userData, celix_bundle_context_t *co
 	struct activator *activator = userData;
 	const char *uuid = NULL;
 
-	status = bundleContext_getProperty(context, OSGI_FRAMEWORK_FRAMEWORK_UUID, &uuid);
+	status = bundleContext_getProperty(context, CELIX_FRAMEWORK_FRAMEWORK_UUID, &uuid);
 	if (!uuid) {
         celix_logHelper_debug(activator->loghelper, "no framework UUID defined?!");
 		return CELIX_ILLEGAL_STATE;