+++ libs/framework/gtest/src/DependencyManagerTestSuite.cc
@@ -35,8 +35,8 @@ public:
     DependencyManagerTestSuite() {
         properties = celix_properties_create();
         celix_properties_set(properties, "LOGHELPER_ENABLE_STDOUT_FALLBACK", "true");
-        celix_properties_set(properties, "org.osgi.framework.storage.clean", "onFirstInit");
-        celix_properties_set(properties, "org.osgi.framework.storage", ".cacheBundleContextTestFramework");
+        celix_properties_set(properties, CELIX_FRAMEWORK_CLEAN_CACHE_DIR_ON_CREATE, "onFirstInit");
+        celix_properties_set(properties, CELIX_FRAMEWORK_CACHE_DIR, ".cacheBundleContextTestFramework");
         celix_properties_set(properties, "CELIX_LOGGING_DEFAULT_ACTIVE_LOG_LEVEL", "trace");