+++ libs/framework/src/bundle_archive.c
@@ -216,13 +216,13 @@ static celix_status_t celix_bundleArchive_createCacheDirectory(bundle_archive_pt
     }
 
     //populate bundle symbolic name and version from manifest
-    archive->bundleSymbolicName = celix_utils_strdup(manifest_getValue(*manifestOut, OSGI_FRAMEWORK_BUNDLE_SYMBOLICNAME));
+    archive->bundleSymbolicName = celix_utils_strdup(manifest_getValue(*manifestOut, CELIX_FRAMEWORK_BUNDLE_SYMBOLICNAME));
     if (archive->bundleSymbolicName == NULL) {
         fw_log(archive->fw->logger, CELIX_LOG_LEVEL_ERROR, "Failed to initialize archive. Cannot read bundle symbolic name.");
         manifest_destroy(*manifestOut);
         return CELIX_BUNDLE_EXCEPTION;
     }
-    archive->bundleVersion = celix_utils_strdup(manifest_getValue(*manifestOut, OSGI_FRAMEWORK_BUNDLE_VERSION));
+    archive->bundleVersion = celix_utils_strdup(manifest_getValue(*manifestOut, CELIX_FRAMEWORK_BUNDLE_VERSION));
     if (archive->bundleVersion == NULL) {
         fw_log(archive->fw->logger, CELIX_LOG_LEVEL_ERROR, "Failed to initialize archive. Cannot read bundle version.");
         manifest_destroy(*manifestOut);