+++ CHANGES.md
@@ -33,6 +33,10 @@ limitations under the License.
 - pubsub_serializer.h is removed and no longer supported. Use pubsub_message_serialization_service.h instead.
 - C++11 support for dm is removed. C++14 is now the minimum required version.
 - C++17 string_view support is removed from the utils and framework lib.
+- Rename of `CELIX_FRAMEWORK_FRAMEWORK_CACHE_DIR` config property to `CELIX_FRAMEWORK_CACHE_DIR`.
+- Support for OSGI_FRAMEWORK_* config properties are dropped. Use CELIX_FRAMEWORK_* instead. 
+  Note this includes the `OSGI_FRAMEWORK_FRAMEWORK_STORAGE` ("org.osgi.framework.storage") config property, 
+  which has been replaced with the `CELIX_FRAMEWORK_CACHE_DIR` config property.
 
 # Noteworthy Changes for 2.4.0 (2023-09-27)